Make BindException mandatory only when needed.
Several commands have no need for the BindException field.  This was
adding a hassle for the web tier, as it meant that a useless
BindException instance was created and then thrown away.  Made a
function within the abstract command that by default says the field
is required, but it can be overridden to not require it for
commands that don't need it.