configured POM to allow integration tests to be run from eclipse
Essentially, this involves moving all of the automated integration setup
that would happen during a build to a profile that is active by default.
This profile is the "automated-integration-tests" profile in the POM.
Additionally, an "ide-integration-tests" profile is added to the POM.
It sets values for all of the properties used for integration testing.
These properties get filtered into test resources in the usual way.  As
long as the "ide-integration-tests" profile is enabled, the
"automated-integration-tests" profile will disable itself.

See the UAlerts wiki for instructions on running integration tests
within Eclipse.